# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "學生選課系統",
        "description": "Laravel 10 學生選課系統 API 文件",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://laravel10-course.test/",
            "description": "本地開發環境"
        }
    ],
    "paths": {
...
        "/api/teachers": {
            "get": {
                "tags": [
                    "教師"
                ],
                "summary": "教師列表",
                "operationId": "9115585a972eee31c1bb78305dc1620c",
                "responses": {
                    "200": {
                        "description": "教師列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Eric"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "test@example.com"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "Eric",
                                                    "email": "test@example.com"
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "Mary",
                                                    "email": "mary@example.com"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到教師",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "找不到教師"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "教師"
                ],
                "summary": "新增教師",
                "operationId": "232f4cec049f7f25ccdce549242471bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "教師姓名",
                                        "type": "string",
                                        "example": "Eric"
                                    },
                                    "username": {
                                        "description": "登入帳號",
                                        "type": "string",
                                        "example": "eric"
                                    },
                                    "password": {
                                        "description": "登入密碼",
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "教師新增成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Eric"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "test@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "姓名是必填的"
                                                    }
                                                },
                                                "username": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "用戶名稱是必填的",
                                                            "用戶名稱已存在"
                                                        ]
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "密碼至少需要8個字元"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "新增教師失敗"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/teachers/{user}/courses": {
            "get": {
                "tags": [
                    "教師"
                ],
                "summary": "教師課程列表",
                "operationId": "c9dd695ab2be52e701eb3e74f7b18850",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "教師ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "課程列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "中級英文"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "中級英文課程，適合初學者"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "0900"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "1030"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "中級英文",
                                                    "description": "中級英文課程，適合初學者",
                                                    "start_time": "0900",
                                                    "end_time": "1030"
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "程式設計入門",
                                                    "description": "學習基礎程式設計概念",
                                                    "start_time": "1400",
                                                    "end_time": "1530"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到教師",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "找不到教師"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}